Feat: 전체 기능 로그인 적용

+++ src/main/java/com/example/todaysbook/controller/RecommendListController.java
@@ -1,12 +1,15 @@
 package com.example.todaysbook.controller;
 
+import com.example.todaysbook.domain.dto.CustomUserDetails;
 import com.example.todaysbook.domain.dto.RecommendListCreateRequestDto;
 import com.example.todaysbook.domain.dto.RecommendListUpdateRequestDto;
 import com.example.todaysbook.domain.dto.RecommendListDetailDto;
 import com.example.todaysbook.domain.entity.UserRecommendList;
+import com.example.todaysbook.exception.user.UserValidateException;
 import com.example.todaysbook.service.RecommendListService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -35,17 +38,16 @@ public class RecommendListController {
 
     @GetMapping("/add")
     public String getWriteForm(Model model) {
-        long loginUserId=1;
 
-        model.addAttribute("login_user_id", loginUserId);
         model.addAttribute("recommendList", new RecommendListCreateRequestDto());
 
         return "user/mypage/create-recommendlist";
     }
 
     @PostMapping("/add")
-    public ResponseEntity<String> createRecommnedList(RecommendListCreateRequestDto request) {
-        long userId = 1;
+    public ResponseEntity<String> createRecommnedList(@AuthenticationPrincipal CustomUserDetails userDetails, RecommendListCreateRequestDto request) {
+
+        Long userId = userDetails.getUserId();
 
         UserRecommendList save = recommendListService.save(userId, request);
 
@@ -54,11 +56,9 @@ public class RecommendListController {
 
     @GetMapping("/update/{id}")
     public String getUpdateForm(@PathVariable Long id, Model model) {
-        long loginUserId=1;
 
         RecommendListDetailDto recommendListDetail = recommendListService.getRecommendListDetail(id);
 
-        model.addAttribute("login_user_id", loginUserId);
         model.addAttribute("recommendList", recommendListDetail);
 
         return "user/mypage/update-recommendlist";
@@ -73,7 +73,14 @@ public class RecommendListController {
     }
 
     @DeleteMapping("/remove/{id}")
-    public ResponseEntity<String> deleteRecommendList(@PathVariable(value = "id") Long listId, Long loginUserId) {
+    public ResponseEntity<String> deleteRecommendList(@AuthenticationPrincipal CustomUserDetails userDetails,
+                                                      @PathVariable(value = "id") Long listId,
+                                                      Long userId) {
+
+        //현재 로그인한 유저랑 리스트 만든 유저랑 다를때 예외 처리
+        if(userDetails.getUserId()!=userId){
+            throw new UserValidateException();
+        }
 
         recommendListService.delete(listId);
